split poll-ingest-job.py into two files, new one being get-job-result.py so that the poll-ingest-job can be used regardless of whether or not we want to get the job result (more modular, should be useful for table ingest ticket), added a result_field option for the get-job-result.py script
